+++ main.rb
@@ -1,38 +1,4 @@
-require_relative 'nameable'
 require_relative 'app'
-
-# Base decorator
-class Decorator < Nameable
-  def initialize(nameable)
-    super()
-    @nameable = nameable
-  end
-
-  attr_accessor :nameable
-
-  def correct_name
-    @nameable.correct_name
-  end
-end
-
-# Capitalize Decorator
-class CapitalizeDecorator < Decorator
-  def correct_name
-    @nameable.correct_name.capitalize
-  end
-end
-
-# Trimmer Decorator
-class TrimmerDecorator < Decorator
-  def correct_name
-    original_name = @nameable.correct_name
-    if original_name.length > 10
-      original_name[0..9] # Trim the name to maximum 10 characters
-    else
-      original_name
-    end
-  end
-end
 
 def display_menu
   puts 'Please choose an option by enterin a number:'
